+++ Assets/NoesisGUI/Xaml/HUD.xaml
@@ -277,6 +277,7 @@
        
     </Grid>
 
+    <!--ScoreBoardGrid-->
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="1*" />
@@ -290,87 +291,6 @@
             <RowDefinition Height="1*" />
         </Grid.RowDefinitions>
 
-        <Viewbox
-                Grid.Column="1"
-                Grid.Row="1">
-
-            <Grid
-                x:Name="SCOREBOARD_GRID"
-                Background="#9B101010"
-                MinHeight="800"
-                MinWidth="1600"
-                Visibility="Hidden">
-
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="1*" />
-                    <ColumnDefinition Width="1*" />
-                </Grid.ColumnDefinitions>
-
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="2*" />
-                    <RowDefinition Height="1*" />
-                </Grid.RowDefinitions>
-
-                <Grid.Resources>
-                </Grid.Resources>
-
-                <StackPanel x:Name="BLUE_TEAM_STACK_PANEL" Grid.Column="0" Margin="30 50 30 50" TextElement.FontSize="20">
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="2*" />
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="1*" />
-                            <ColumnDefinition Width="1*" />
-                            <ColumnDefinition Width="1*" />
-                        </Grid.ColumnDefinitions>
-
-                        <Label Content="Blue Team" Grid.Column="1" Foreground="#05DFD7" FontSize="30" VerticalAlignment="Bottom"/>
-                        <Label Content="K" Grid.Column="2" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="D" Grid.Column="3" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Flags cap." Grid.Column="4" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Flags def." Grid.Column="5" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Ping" Grid.Column="6" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                    </Grid>
-
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="6*" />
-                        </Grid.ColumnDefinitions>
-                        <Separator Grid.Column="1" Background="White" />
-                    </Grid>
-                </StackPanel>
-
-                <StackPanel x:Name="RED_TEAM_STACK_PANEL" Grid.Column="1" Margin="50" TextElement.FontSize="20">
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="2*" />
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="1*" />
-                            <ColumnDefinition Width="1*" />
-                            <ColumnDefinition Width="1*" />
-                        </Grid.ColumnDefinitions>
-
-                        <Label Content="Red Team" Grid.Column="1" Foreground="#E52C2C" FontSize="30" VerticalAlignment="Bottom"/>
-                        <Label Content="K" Grid.Column="2" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="D" Grid.Column="3" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Flags cap." Grid.Column="4" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Flags def." Grid.Column="5" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                        <Label Content="Ping" Grid.Column="6" Foreground="White" HorizontalAlignment="Right" VerticalAlignment="Bottom"/>
-                    </Grid>
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="0.5*" />
-                            <ColumnDefinition Width="6*" />
-                        </Grid.ColumnDefinitions>
-                        <Separator Grid.Column="1" Background="White" />
-                    </Grid>
-                </StackPanel>
-            </Grid>
-        </Viewbox>
+       <uc:ScoreBoardGUI x:Name="SCORE_BOARD_GUI" HorizontalAlignment="Center" VerticalAlignment="Center"  Grid.Column="1" Grid.Row="1"/>
     </Grid>
 </Grid>
