+++ CrazyCanvas/Include/GUI/HUDGUI.h
@@ -30,6 +30,7 @@
 
 #include "GUI/EscapeMenuGUI.h"
 #include "GUI/KillFeedGUI.h"
+#include "GUI/ScoreBoardGUI.h"
 
 #include "NsCore/BaseComponent.h"
 #include "NsCore/Type.h"
@@ -50,16 +51,6 @@ struct GameGUIState
 
 	int32 Ammo;
 	int32 AmmoCapacity;
-};
-
-enum class EPlayerProperty
-{
-	PLAYER_PROPERTY_NAME,
-	PLAYER_PROPERTY_KILLS,
-	PLAYER_PROPERTY_DEATHS,
-	PLAYER_PROPERTY_FLAGS_CAPTURED,
-	PLAYER_PROPERTY_FLAGS_DEFENDED,
-	PLAYER_PROPERTY_PING,
 };
 
 typedef  std::pair<uint8, const Player*> PlayerPair;
@@ -84,11 +75,6 @@ public:
 	void DisplayGameOverGrid(uint8 winningTeamIndex, PlayerPair& mostKills, PlayerPair& mostDeaths, PlayerPair& mostFlags);
 	void DisplayPrompt(const LambdaEngine::String& promptMessage);
 
-	void AddPlayer(const Player& newPlayer);
-	void RemovePlayer(const Player& player);
-	void UpdatePlayerProperty(uint64 playerUID, EPlayerProperty property, const LambdaEngine::String& value);
-	void UpdateAllPlayerProperties(const Player& player);
-	void UpdatePlayerAliveStatus(uint64 UID, bool isAlive);
 	void UpdateKillFeed(const LambdaEngine::String& killed, const LambdaEngine::String& killer, uint8 killedPlayerTeamIndex);
 	void UpdateKillFeedTimer(LambdaEngine::Timestamp delta);
 
@@ -98,15 +84,14 @@ public:
 
 	void SetWindowSize(uint32 width, uint32 height);
 
+	ScoreBoardGUI* GetScoreBoard() const;
+
 private:
 	void InitGUI();
 
 	void TranslateIndicator(Noesis::Transform* pTranslation, LambdaEngine::Entity entity);
 	void SetIndicatorOpacity(float32 value, LambdaEngine::Entity entity);
 	void SetRenderStagesInactive();
-
-	// Helpers
-	void AddStatsLabel(Noesis::Grid* pParentGrid, const LambdaEngine::String& content, uint32 column);
 
 	NS_IMPLEMENT_INLINE_REFLECTION_(HUDGUI, Noesis::Grid)
 
@@ -121,7 +106,6 @@ private:
 	Noesis::TextBlock* m_pWaterAmmoText			= nullptr;
 	Noesis::TextBlock* m_pPaintAmmoText			= nullptr;
 
-
 	Noesis::Grid* m_pHUDGrid					= nullptr;
 
 	Noesis::Grid* m_pHitIndicatorGrid			= nullptr;
@@ -129,9 +113,6 @@ private:
 
 	Noesis::Grid* m_pRedScoreGrid				= nullptr;
 	Noesis::Grid* m_pBlueScoreGrid				= nullptr;
-
-	Noesis::StackPanel* m_pBlueTeamStackPanel	= nullptr;
-	Noesis::StackPanel* m_pRedTeamStackPanel	= nullptr;
 
 
 	glm::vec2 m_WindowSize = glm::vec2(1.0f);
@@ -144,5 +125,5 @@ private:
 
 	KillFeedGUI* m_pKillFeedGUI		= nullptr;
 	EscapeMenuGUI* m_pEscMenuGUI	= nullptr;
-
+	ScoreBoardGUI* m_pScoreBoardGUI = nullptr;
 };
