+++ CrazyCanvas/Saved_Servers.json
@@ -1,3 +1,8 @@
 {
-    "SERVERS": []
+    "SERVERS": [
+        {
+            "Name": "",
+            "Address": "192.168.1.65:4444"
+        }
+    ]
 }
